Disable auto-updates and update project metadata
The `<UpdateEnabled>` property was set to `false` and `<UpdateInterval>` to `0`, disabling automatic updates. Updated `<PublisherName>` to "Andy Gock" and `<SupportUrl>` to "https://github.com/andygock/SendGuard". Added a detailed `<OfficeApplicationDescription>`.

Replaced the temporary signing key file with `AndyGock.pfx` and updated the `<ManifestCertificateThumbprint>` to reflect the new certificate. Added `AndyGock.pfx` to the project.